SSDI final Project- WebApp
This project consists of mobile and WebApp extracting information from a database hosted on Azure. I m responsible for the entire WebApp. since it involves costs for hosting database in azure, database instance is terminated after project presentation. The WebApp makes AJAX calls to the API and displays data, so for now , u can look at the code.