course: Network based Application Development
This is a webApp that gets its data from RDS instance hosted on AWS, which is connected to a MySQL server that will run on my laptop while navigating through the pages. Since the database instance is turned down after project presentation, you can view the code for this project.
built with : JavaScript,bootstrap3, MVC architecture .
concepts used :

creating a MySQL database for the intended functions of webApp. Storing, retrieving and editing the data in tables by user input and DOM manipulation

Hashing and salting passwords of user data

To create an RDS instance on AWS and connecting it to my local MySQL server

How to use "JavaMail" feature